requirements.txt
pyTelegramBotAPI
unstructured
faiss-cpu
langchain
langchain-openai
python-dotenv